Refactor: Renamed homepage components
The src/components/Banner.tsx and src/components/Header.tsx were
renamed to src/components/HomepageBanner.tsx and
src/components/HomepageHeader.tsx to better reflect the fact that they
will only be used at the homepage.

This change was also made so that they will not conflict with the Banner
and Header for the continents pages that will be added further down the
road.